+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project follows [Semantic Versioning](http://semver.org/).
 ### Changed
 - Updated variable title to display Form or Item Variables.
 - Updated empty variable list message.
+- Added support for enableWhen expression.
 
 ## [3.3.2] 2025-08-26
 ### Fixed